+++ src/error.rs
@@ -1,4 +1,5 @@
 use std::sync::Arc;
+use jsonwebtoken::errors::ErrorKind;
 
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
@@ -27,5 +28,20 @@ impl From<Error> for JwksClientError {
 impl From<jsonwebtoken::errors::Error> for JwksClientError {
     fn from(error: jsonwebtoken::errors::Error) -> Self {
         Self::Error(Arc::new(error.into()))
+    }
+}
+
+impl JwksClientError {
+    pub fn is_jwt_expired(&self) -> bool {
+        match self { JwksClientError::Error(e) => { e.is_jwt_expired() } }
+    }
+}
+
+impl Error {
+    fn is_jwt_expired(&self) -> bool {
+        match self {
+            Error::JsonWebToken(err) => { matches!(err.kind(), ErrorKind::ExpiredSignature) }
+            _ => false
+        }
     }
 }